Added node.js based QIDO /WADO api and simple cornerstone based api
Luiz Oliveira worked very hard to create these new options, and I helped by proving fast access to json utilities in the dicom server (dgate binary it must be up to date for it to work). This is a first test release. To see how to install and run it see the readme files in the api and app folders.